% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_args_here.R
\name{all_args_here}
\alias{all_args_here}
\title{All Arguments Here}
\usage{
all_args_here(
  x,
  export = FALSE,
  out_path = ".",
  df_name = "res.tsv",
  overwrite = FALSE,
  safer_check = TRUE,
  lib_path = NULL,
  error_text = ""
)
}
\arguments{
\item{x}{Function name, written without quotes and brackets.}

\item{export}{Single logical value. Export the result data frame into a \code{.tsv} file? If \code{TRUE}, the data frame is not returned by the function but only exported.}

\item{out_path}{Single character string indicating the absolute pathway of the folder where to export the data frame. \code{out_path = "."} means the R working directory set by the user. Ignored if export is \code{FALSE}.}

\item{df_name}{Single character string indicating the name of the exported data frame file. Ignored if export is \code{FALSE}.}

\item{overwrite}{Single logical value. Overwrite potential \code{df_name} file already existing in \code{out_path}? Ignored if export is \code{FALSE}.}

\item{safer_check}{Single logical value. Perform the "safer" checks? If \code{TRUE}, checkings are performed before main code running (see the \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Warning: must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{c(lib_path, .libPaths())}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
A data frame indicating the missing arguments or a message saying that everything seems fine.
If \code{export} argument is \code{TRUE}, then the data frame is exported as \code{res.tsv} instead of being returned.
Column descriptions: 
\itemize{
  \item \code{$LINE_NB}: the line number in the function code (starting at the \code{"<- function"} line, i.e., without counting the \code{#'} header lines).
  \item \code{$FUN_NAME}: the function name.
  \item \code{$FUN_ARGS}: the written arguments of \code{FUN_NAME}. \code{"NOT_CONSIDERED"} means that the function is between quotes or after \code{$}.
  \item \code{$FUN_POS}: the position of the first character of the function name in the \code{$LINE_NB} line.
  \item \code{$DEF_ARGS}: the defaults arguments of \code{FUN_NAME}. \code{"NO_ARGS"} means that the function has no arguments. \code{"INTERNAL_FUNCTION"} means that the function has been created inside the checked function. \code{"SKIPPED"} means that the function is not analyzed for the reason indicated in the "details" section.
  \item \code{$MISSING_ARG_NAMES}: the missing argument names in \code{FUN_ARGS}.
  \item \code{$MISSING_ARGS}: the missing arguments with their values in \code{FUN_ARGS}.
  \item \code{$STATUS}: either \code{"GOOD"}, meaning that all the arguments are already written, or a new proposal of arguments writing, or indicates if some arguments are not fully written (abbreviation is discouraged), or nothing.
}

An additional message \code{"EVERYTHING SEEMS CLEAN"} if the \code{STATUS} column is only made of \code{""} and \code{"GOOD"}.
}
\description{
Verify that all the functions used inside a function are written with all their arguments. For instance: \code{base::paste0(letters[1:2], collapse = NULL, recycle0 = FALSE)} and not \code{paste0(letters[1:2])}.
}
\details{
More precisely, \code{all_args_here()} verifies that all the strings before an opening bracket \code{(} are written with all their arguments. Thus, it cannot check function names written without brackets, like in the \code{FUN} argument of some functions, e.g., \code{sapply(1:3, FUN = as.character)}.

The Perl regex used to detect a function name is: \code{"([a-zA-Z]|\\.[a-zA-Z._])[a-zA-Z0-9._]*\\s*\\("}.

Currently, \code{all_args_here()} cannot detect functions written between quotes, like \code{"+"()} or \code{"rownames<-"(x, "a")}.

Function names preceded by \code{$} are not considered.
 
The following R functions are skipped: \code{function}, \code{if}, \code{for}, \code{while}, \code{repeat} and \code{else}.

Most of the time, \code{all_args_here()} does not check inside comments, but some unexpected writing could dupe \code{all_args_here()}. Please, report \href{https://github.com/safer-r/saferDev/issues}{here} if it is the case.

The returned line numbers are indicative, depending on which source is checked. For instance, \code{saferDev::report} (compiled) has not the same line numbers as its \href{https://github.com/safer-r/saferDev/blob/main/R/report.R}{source file}. Notably, compiled functions do not have comments anymore, compared to the same source function sourced into the working environment. In addition, the counting starts at the \code{"<- function"} line, i.e., without counting the \code{#'} header lines potentially present in source files.

The function works first by replacing in the code: 1) the three consecutive characters \code{'"'} or \code{'"'} by three spaces, 2) \code{"\"'"} and \code{'\'"'} by four spaces and 3) escape quotes like \code{\"} or \code{\'} by two spaces.

See more examples \href{https://safer-r.github.io/saferDev/articles/all_args_here.html}{here}.

Warnings:
\enumerate{
  \item The following R functions are also skipped (as indicated by \code{"SKIPPED"} in the \code{DEF_ARGS} column of the returned data frame): \href{https://bugs.r-project.org/show_bug.cgi?id=18849}{\code{as.environment()}}. Click to see the explanation.
  \item Some functions, like \code{rownames()}, have different arguments depending on whether something is assigned to it (e.g., \code{rownames(x) <- "a"}) or not. The \code{all_args_here()} function always proposes the arguments defined in the help page without assignment, meaning that it cannot detect the assignment. A way to bypass this is to use the "exact" writing of the function. For instance, \code{base::"rownames<-"(x, "a")} instead of using \code{base::rownames(x) <- "a"}, and use \code{getAnywhere("rownames<-")} to see the arguments of the function.
  \item The function could not properly work if any comma is present in default argument values. Please, report \href{https://github.com/safer-r/saferDev/issues}{here} if it is the case.
  \item Proposals in the \code{STATUS} column are only suggestions, as it is difficult to anticipate all the exceptions with arguments writing.
}
}
\examples{
# Warning: these examples may not work well when using the "Run examples" link 
# because of a particular environment. Please, copy-paste in a local environment.
# See also https://safer-r.github.io/saferDev/articles/all_args_here.html
\dontrun{
# Example that returns an error
saferDev::all_args_here(mean) # Example that returns an error
source("https://raw.githubusercontent.com/safer-r/saferDev/main/dev/other/test2.R")
saferDev::all_args_here(test2) # the checked function must be executable
}

FUN2 <- function(x, y){middle_bracket2 <- base::do.call(what = base::c, args = code_for_col, quote = FALSE, envir = base::parent.frame())}
saferDev::all_args_here(FUN2, safer_check = FALSE)

\dontrun{
# Example that creates a file/folder in the working directory
source("https://raw.githubusercontent.com/safer-r/.github/refs/heads/main/profile/backbone.R")
saferDev::all_args_here(BACKBONE, export = TRUE, safer_check = FALSE)
}

}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

\href{yushi.han2000@gmail.com}{Yushi Han}

\href{wanghaiding442@gmail.com}{Haiding Wang}
}
