\name{cuteDev-package}
\alias{cuteDev-package}
\alias{cuteDev}
\docType{package}
\title{
\packageTitle{cuteDev}
}
\description{
\packageDescription{cuteDev}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cuteDev}
\packageIndices{cuteDev}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cuteDev}

Maintainer: \packageMaintainer{cuteDev}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
