% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.colons_check_message.R
\name{.colons_check_message}
\alias{.colons_check_message}
\title{Colons Check Message}
\usage{
.colons_check_message(
  list_fun,
  list_fun_pos,
  line_nb,
  ini,
  arg_user_setting2,
  text,
  internal_fun_names,
  lib_path,
  error_text
)
}
\arguments{
\item{list_fun}{List of names of all the functions.}

\item{list_fun_pos}{List of positions of first character of names of all the functions in \code{ini}.}

\item{line_nb}{Vector of integers of corresponding line numbers.}

\item{ini}{Vector of strings of the initial function code analyzed.}

\item{arg_user_setting2}{List of argument user settings.}

\item{text}{Either \code{"BASIC"} or \code{"OTHER"}.}

\item{internal_fun_names}{Vector of strings of names of internal functions in the function code analyzed. Can be \code{NULL}.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
A list:
\itemize{
  \item \code{output.cat}: the message (string).
  \item \code{colon_not_here}: logical vector. Does \code{list_fun} contain function names without \code{::} or \code{:::} ?
}
}
\description{
Create the message for the \code{colons_check()} function.
}
\details{
Warnings:
\itemize{
  \item requires \code{saferDev::arg_check}, \code{.noclean_functions}.
  \item In main safer functions, in the section \code{"######## check of the required functions from the required packages"}, add also these 2 functions when checking for the presence of \code{.colons_check_message}.
}
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

Haiding Wang

Yushi Han
}
\keyword{internal}
