% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.extract_all_fun_names.R
\name{.extract_all_fun_names}
\alias{.extract_all_fun_names}
\title{Extract All Function Names}
\usage{
.extract_all_fun_names(text, pattern, lib_path, error_text)
}
\arguments{
\item{text}{Single string.}

\item{pattern}{Single string of a Perl regex to extract function names.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
List containing:
\itemize{
  \item \code{string}: the function names without parenthesis.
  \item \code{pos}: position of the first character of the function names in the input string
}
}
\description{
Extract all function names.
}
\details{
Warnings:
\itemize{
  \item requires \code{saferDev::arg_check}.
  \item In the safer Backbone section \code{"######## check of the required functions from the required packages"}, add also \code{saferDev::arg_check} when checking for the presence of \code{.extract_all_fun_names}.
}
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

Haiding Wang

Yushi Han
}
\keyword{internal}
