% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.has_odd_number_of_quotes.R
\name{.has_odd_number_of_quotes}
\alias{.has_odd_number_of_quotes}
\title{Has Odd Number of Quotes}
\usage{
.has_odd_number_of_quotes(input_string, pattern, lib_path, error_text)
}
\arguments{
\item{input_string}{Single string.}

\item{pattern}{Either \code{'"'} or \code{"'"}.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Determine if a string has a odd number of quotes.
}
\details{
The function considers any present \verb{'} (\code{pattern = "'\''"}) or \verb{"} (\code{pattern = '"'}). Meaning that it does not treat special situation, like escape quotes \code{\"}, or situation like this \code{"'"}. All these must be removed from the \code{input_string} if necessary, using for instance: \code{input_string <- gsub( x = input_string, pattern = '"\'"', replacement = '   ')}

Warnings:
\itemize{
  \item requires \code{saferDev::arg_check}.
  \item In the safer Backbone section \code{"######## check of the required functions from the required packages"}, add also \code{saferDev::arg_check} when checking for the presence of \code{saferDev:::.has_odd_number_of_quotes}.
}
}
\examples{
\dontrun{
# Example that shouldn't be run because this is an internal function.
saferDev:::.has_odd_number_of_quotes(
input_string = 'This is a "test" string with "even" quotes', 
pattern = '"', 
lib_path = NULL, 
error_text = " INSIDE P1::F1"
)
}


}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

Haiding Wang

Yushi Han
}
\keyword{internal}
