% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.noclean_functions.R
\name{.noclean_functions}
\alias{.noclean_functions}
\title{No-Clean Functions}
\usage{
.noclean_functions(col1, col2, col3, ini, lib_path, error_text)
}
\arguments{
\item{col1}{Vector of integers of line numbers in a code.}

\item{col2}{Vector of strings of the function names scrutinized.}

\item{col3}{Vector of strings of the code before the function name.}

\item{ini}{Vector of strings of the initial function code analyzed.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
A logical vector indicating if function names of \code{col2} are inside quotes or after \code{$} (\code{TRUE}) in \code{ini} or not (\code{FALSE}). Can be length 0.
}
\description{
Indicate if function names are inside quotes or after \code{$}.
}
\details{
Warnings:
\itemize{
  \item requires \code{saferDev::arg_check}, \code{.has_odd_number_of_quotes}.
  \item In the safer Backbone section \code{"######## check of the required functions from the required packages"}, add also these 2 functions when checking for the presence of \code{.noclean_functions}.
}
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

Haiding Wang

Yushi Han
}
\keyword{internal}
