% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{fun_pack}
\usage{
fun_pack(req.package, load = FALSE, lib.path = NULL)
}
\arguments{
\item{req.package}{Character vector of package names to import.}

\item{load}{Logical. Load the package into the environment (using library())? Interesting if packages are not in default folders or for checking the functions names of packages using search().}

\item{lib.path}{Optional character vector specifying the absolute pathways of the directories containing some of the listed packages in the req.package argument, if not in the default directories. Ignored if NULL.}
}
\value{
Nothing.
}
\description{
Check if the specified R packages are present in the computer and import them into the working environment.
}
\details{
REQUIRED PACKAGES

none

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

fun_check()
}
\examples{
# fun_pack(req.package = "nopackage")
fun_pack(req.package = "ggplot2")
fun_pack(req.package = "ggplot2", lib.path = "C:/Users/yhan/AppData/Local/R/win-library/4.3")
}
