% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_report.R
\name{fun_report}
\alias{fun_report}
\title{fun_report}
\usage{
fun_report(
  data,
  output = "results.txt",
  path,
  overwrite = FALSE,
  rownames.kept = FALSE,
  vector.cat = FALSE,
  noquote = TRUE,
  sep = 2
)
}
\arguments{
\item{data}{Object to print in the output file. If NULL, nothing is done, with no warning.}

\item{output}{Name of the output file.}

\item{path}{Location of the output file.}

\item{overwrite}{(logical) if output file already exists, defines if the printing is appended (default FALSE) or if the output file content is erased before printing (TRUE).}

\item{rownames.kept}{(logical) defines whether row names have to be removed or not in small tables (less than length.rows rows).}

\item{vector.cat}{(logical). If TRUE print a vector of length > 1 using cat() instead of capture.output(). Otherwise (default FALSE) the opposite.}

\item{noquote}{(logical). If TRUE no quote are present for the characters.}

\item{sep}{Number of separating lines after printed data (must be integer).}
}
\value{
Nothing.
}
\description{
Log file function: print a character string or a data object into a same output file.
}
\details{
REQUIRED PACKAGES

none

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

fun_check()
}
\examples{
#fun_report()
fun_report(data = 1:3, output = "results.txt", path = "C:/Users/yhan/Desktop", overwrite = TRUE, 
rownames.kept = FALSE, vector.cat = FALSE, noquote = FALSE, sep = 2)
}
