% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message.R
\name{get_message}
\alias{get_message}
\title{Get Message}
\usage{
get_message(
  data,
  kind = "error",
  header = TRUE,
  print_no = FALSE,
  text = NULL,
  env = NULL,
  safer_check = TRUE,
  lib_path = NULL,
  error_text = ""
)
}
\arguments{
\item{data}{Single character string of an instruction to evaluate}

\item{kind}{Single character string. Either \code{"error"} to get a potential error message, or \code{"warning"} to get a potential warning message, or \code{"message"} to get a potential standard (non error and non warning) message.}

\item{header}{Single logical value. Add a header in the returned message?}

\item{print_no}{Single logical value. Print a message saying that nothing (\code{NULL} output) has to be reported?}

\item{text}{Single character string added to the header of the output message, even if nothing reported (\code{print_no} is \code{TRUE}). Ignored if the \code{header} argument is \code{FALSE}. Write \code{NULL} if not required.}

\item{env}{An object corresponding to an existing environment. Then the \code{data} argument value is evaluated only in the indicated environment. Write \code{NULL} if not required (R scope is used). Example \code{env = .GlobalEnv}. Example \code{env = asNamespace("stats")}.}

\item{safer_check}{Single logical value. Perform some "safer" checks? If \code{TRUE}, checkings are performed before main code running (see the \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
The function returns the error, warning, or standard message, as a single character string, depending on what has been selected using the \code{kind} argument and if such message exists.

\code{NULL} if no selected message returned and the \code{print_no} argument is \code{FALSE}.

The following message if no selected message returned and the \code{print_no} argument is \code{TRUE}: \code{"NO (ERROR|WARNING|STANDARD) MESSAGE REPORTED"}.

The following message if 1) the \code{kind} argument is \code{"warning"} or \code{"message"} while an error message is returned and 2) the \code{print_no} argument is \code{TRUE}: \code{"NO POTENTIAL (WARNING|STANDARD) MESSAGE BECAUSE OF ERROR MESSAGE REPORTED"}.
}
\description{
Evaluate an instruction written between \code{""} and return the first of the error message, or the last of the warning or standard (non error non warning) messages if ever exist.

Using argument \code{print_no = FALSE}, return \code{NULL} if nothing reported, which is convenient in some cases.
}
\details{
Warnings:
\itemize{
  \item Only the first error or last warning/standard message is returned.
  \item Always use the \code{env} argument when \code{get_message()} is used inside functions.
  \item The function does not prevent printing, for instance if \code{print()} or \code{show()} is used inside the instruction tested. To prevent that, use \code{tempo <- utils::capture.output(error <- get_message(data = "arg_check(data = 'a', class = mean, neg_values = FALSE, print = TRUE)"))}. The return of \code{get_message()} is assigned into \code{error} and the printed messages are captured by \code{utils::capture.output()} and assigned into \code{tempo}. See the examples.
}

The \code{env} argument is used as value for the \code{envir} argument of the \code{base::eval()} function.
}
\examples{
# Warning: these examples may not work well when using the "Run examples" link 
# because of a particular environment. Please, copy-paste in a local environment.
# See also https://safer-r.github.io/saferDev/articles/get_message.html

# Report error message by default. Here no error to report:
saferDev::get_message(data = "wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)")

# Warning message to report:
saferDev::get_message(data = "wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)", kind = "warning")

# No standard message to report:
saferDev::get_message(data = "wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)", kind = "message", 
print_no = TRUE, text = "IN A")

# Standard message cannot be caught because wilcox.test() returns an error message:
saferDev::get_message(data = "wilcox.test()", kind = "message", print_no = TRUE, text = "IN A")

# Text added in warning message to report:
saferDev::get_message(data = "wilcox.test()", kind = "error", print_no = TRUE, text = "IN A")

# No error message to report with text added:
saferDev::get_message(data = "sum(1)", kind = "error", print_no = TRUE, text = "IN A")

# No error message to report with text added:
saferDev::get_message(data = "message('ahah')", kind = "error", print_no = TRUE, text = "IN A")

# Standard message reported with text added:
saferDev::get_message(data = "message('ahah')", kind = "message", print_no = TRUE, text = "IN A")

# Messages from ggplot2 functions caugth:
saferDev::get_message(data = "ggplot2::ggplot(data = data.frame(X = 1:10, stringsAsFactors = TRUE), 
mapping = ggplot2::aes(x = X)) + ggplot2::geom_histogram()", kind = "message", 
print_no = TRUE, text = "IN INSTRUCTION 1")
}
\seealso{
\code{\link{try}}.
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

Haiding Wang

Yushi Han
}
