% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{pack}
\usage{
pack(req.package, load = FALSE, lib.path = NULL)
}
\arguments{
\item{req.package}{Character vector of package names to import.}

\item{load}{Single logical value. Load the package into the environment (using library())? Interesting if packages are not in default folders or for checking the functions names of packages using search().}

\item{lib.path}{Optional character vector specifying the absolute pathways of the directories containing some of the listed packages in the req.package argument, if not in the default directories. Ignored if NULL.}
}
\value{
Nothing.
}
\description{
Check if the specified R packages are present in the computer and import them into the working environment.
}
\details{
REQUIRED PACKAGES

none


REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

check()
}
\examples{
# pack(req.package = "nopackage") # should return an error
pack(req.package = "ggplot2")
pack(req.package = "ggplot2", lib.path = "C:/Users/yhan/AppData/Local/R/win-library/4.3") 
# should return an error if the lib.path argument is not an existing directory
}
