% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(
  data,
  output = "log.txt",
  path,
  overwrite = FALSE,
  rownames_kept = FALSE,
  vector_cat = FALSE,
  noquote = TRUE,
  sep = 2,
  safer_check = TRUE,
  lib_path = NULL,
  error_text = ""
)
}
\arguments{
\item{data}{Object to print in the output file. If \code{NULL}, nothing is done, with no warning.}

\item{output}{Single character string. Name of the output file.}

\item{path}{Single character string indicating the path where to write the output file.}

\item{overwrite}{Single logical value. If the output file already exists and \code{overwrite} is \code{TRUE}, an error message is returned (no overwrite of existing file). Otherwise, the printing is appended (and the output file is created if it does not exist yet).}

\item{rownames_kept}{Single logical value. Defines whether row names have to be removed or not in 2D objects. Warning: in 1D tables, names over the values are taken as row names, and are thus removed if \code{rownames_kept} is \code{FALSE}.}

\item{vector_cat}{Single logical value. If \code{TRUE}, print a vector of length > 1 using \code{cat()} instead of \code{capture.output()}. Otherwise (default \code{FALSE}) print a vector of length > 1 using \code{capture.output()}. Names of values are not printed when \code{TRUE}.}

\item{noquote}{Single logical value. If \code{TRUE} no quote are added to the returned character strings.}

\item{sep}{Single non null and positive integer representing the number of empty lines added in the output file after the printed data.}

\item{safer_check}{Single logical value. Perform some "safer" checks? If \code{TRUE}, checkings are performed before main code running (see the \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
Nothing (implementation of an existing file) or a text file.
}
\description{
Print a character string or a data object into a same output file. Convenient for log file generation.
}
\examples{
\dontrun{
# Example that creates a file/folder in the working directory
saferDev::report(data = "THE FOLLOWING VECTOR IS:\n", output = "results.txt", path = ".", 
overwrite = TRUE, sep = 1)
saferDev::report(data = 1:3, output = "results.txt", path = ".", overwrite = FALSE, 
rownames_kept = FALSE, vector_cat = FALSE, noquote = FALSE, sep = 2)
saferDev::report(data = "THE FOLLOWING MATRIX IS:\n", output = "results.txt", path = ".", 
overwrite = FALSE, sep = 1)
saferDev::report(data = matrix(1:5), output = "results.txt", path = ".", overwrite = FALSE, 
rownames_kept = FALSE, vector_cat = FALSE, noquote = FALSE, sep = 5)
saferDev::report(data = "THE FOLLOWING DATA FRAME IS:\n", output = "results.txt", path = ".", 
overwrite = FALSE, sep = 1)
saferDev::report(data = data.frame(A = 1:8, B = letters[1:8]), output = "results.txt", path = ".", 
overwrite = FALSE, rownames_kept = FALSE, vector_cat = FALSE, noquote = FALSE, sep = 1)
}
}
\seealso{
\code{\link{capture.output}}.
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

\href{mailto:yushi.han2000@gmail.com}{Yushi Han}

\href{mailto:wanghaiding442@gmail.com}{Haiding Wang}
}
