% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_args_here.R
\name{all_args_here}
\alias{all_args_here}
\title{all_args_here}
\usage{
all_args_here(
  x,
  export = FALSE,
  path_out = ".",
  df_name = "res.tsv",
  overwrite = FALSE,
  lib_path = NULL,
  safer_check = TRUE
)
}
\arguments{
\item{x}{a function name, written without quotes and brackets.}

\item{export}{Single logical value. Export the data frame into a .tsv file? If TRUE, the data frame is not returned by the function but only exported.}

\item{path_out}{Single character string indicating the absolute pathway of the folder where to export the data frame. path_out = "." means the R working directory set by the user. Ignored if export is FALSE}

\item{df_name}{Single character string indicating the name of the exported data frame file. Ignored if export is FALSE.}

\item{overwrite}{Single logical value. Overwrite potential df_name file already existing in path_out? Ignored if export is FALSE.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages if not in the default directories. If NULL, the pathway specified by .libPaths() is used.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Must be set to FALSE if all_args_here() fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
A data frame indicating the missing arguments or a message saying that everything seems fine.
If export argument is TRUE, then the data frame is exported as res.tsv instead of being returned.
Data frame: 
 $LINE_NB: the line number in the function code (starting at the "<- function" line, i.e., without counting the #' header lines)
 $FUN_NAME: the function name.
 $FUN_ARGS: the written arguments of $FUN_NAME. "NOT_CONSIDERED" means that the function is between quotes or after $
 $FUN_POS: the position of the first character of the function name in the $LINE_NB line of the code.
 $DEF_ARGS: the defaults arguments of $FUN_NAME. "NO_ARGS" means that the function has no arguments
 $MISSING_ARG_NAMES: the missing argument names in $FUN_ARGS.
 $MISSING_ARGS: the missing arguments with their values in $FUN_ARGS.
 $STATUS: either "GOOD", meaning that all the arguments are already written, or a new proposal of arguments writting, or indicates if some arguments are not fully written (abbreviation is discouraged), or nothing.

An additionnal message "EVERYTHING SEEMS CLEAN" if the $STATUS column is only made of "" and "GOOD".
}
\description{
Verify that all the functions used inside a function are written with all their arguments. For instance: base::paste0(letters[1:2], collapse = NULL, recycle0 = FALSE) and not paste0(letters[1:2]).
}
\details{
More precisely, all_args_here() verifies that all the strings before an opening bracket "(" are written with all their arguments. Thus, it cannot check function names written without brackets, like in the FUN argument of some functions, e.g., sapply(1:3, FUN = as.character).

The perl regex used to detect a function name is: "[a-zA-Z.]{1}[a-zA-Z0-9._]*\\s*\\(".

Function names preceeded by $ and any space are not considered (pattern "\\$ *[a-zA-Z.]{1}[a-zA-Z0-9._]* *\\(")
 
The following R functions using bracket are not considered: "function", "if", "for", "while" and "repeat".

Most of the time, all_args_here() does not check inside comments, but some unexpected writting could dupe all_args_here().

The returned line numbers are indicative, depending on which source is checked. For instance, saferDev::report (compiled) has not the same line numbers as its source file (https://github.com/safer-r/saferDev/blob/main/R/report.R). Notably, compiled functions do not have comments anymore, compared to the same source function sourced into the working environment. In addition, the counting starts at the "<- function" line, i.e., without counting the #' header lines potentially present in source files.

Warnings: 
1) The function could not properly work if any comma is present in default argument values. Please, report here https://github.com/safer-r/saferDev/issues if it is the case.

2) Proposals in the STATUS column are only suggestions, as it is difficult to anticipate all the exceptions with arguments writting.
}
\examples{
\dontrun{ # Example that return an error
all_args_here(mean)
source("https://raw.githubusercontent.com/safer-r/saferDev/main/dev/other/test2.R") ; all_args_here(test2, export = TRUE)
}
FUN2 <- function(x, y){middle_bracket2 <- base::do.call(what = base::c, args = code_for_col, quote = FALSE, envir = base::parent.frame())}
all_args_here(FUN2)
source("https://raw.githubusercontent.com/safer-r/.github/refs/heads/main/profile/backbone.R")
all_args_here(BACKBONE, export = TRUE)
}
\author{
Gael Millot <gael.millot@pasteur.fr>
}
