% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{arg_check}
\alias{arg_check}
\title{Argument Check}
\usage{
arg_check(
  data,
  class = NULL,
  typeof = NULL,
  mode = NULL,
  length = NULL,
  prop = FALSE,
  double_as_integer_allowed = FALSE,
  options = NULL,
  all_options_in_data = FALSE,
  na_contain = FALSE,
  neg_values = TRUE,
  inf_values = TRUE,
  print = FALSE,
  data_name = NULL,
  data_arg = TRUE,
  safer_check = TRUE,
  lib_path = NULL,
  error_text = ""
)
}
\arguments{
\item{data}{Object to test.}

\item{class}{Single character string. Either one of the \code{class()} result or \code{"vector"} or \code{"ggplot2"} (i.e., objects of class \code{c("gg", "ggplot")} for R < 4.5 and \code{c("ggplot2::ggplot", "ggplot", "ggplot2::gg", "S7_object", "gg")} otherwise) or \code{"ggplot_built"} (i.e., objects of class \code{"ggplot_built"} for R < 4.5 and \code{c("ggplot2::ggplot_built", "ggplot_built", "ggplot2::gg", "S7_object")} otherwise) or \code{NULL}. See the warning section below.}

\item{typeof}{Single character string. Either one of the \code{typeof()} result or \code{NULL}.}

\item{mode}{Single character string. Either one of the \code{mode()} result (for non-vector object) or \code{NULL}.}

\item{length}{Single numeric value indicating the length of the object. Not considered if \code{NULL}.}

\item{prop}{Single logical value. Are the numeric values between 0 and 1 (proportion)? If \code{TRUE}, can be used alone, i.e., without checking the class, mode, etc., of the object to test.}

\item{double_as_integer_allowed}{Single logical value. If \code{TRUE}, no error is reported in the checking message if argument is set to \code{typeof = "integer"} or \code{class = "integer"}, while the reality is \code{typeof == "double"} or \code{class == "numeric"} but with zero as modulo (reminder of a division). This means that \code{i <- 1}, which is \code{typeof(i) == "double"} is considered as an integer when setting \code{double_as_integer_allowed = TRUE}. Warning: \code{double_as_integer_allowed = TRUE} uses \code{data \%\% 1 == 0L} but not \code{isTRUE(all.equal(data \%\% 1, 0))} is used here because the argument checks for integers stored as double (does not check for decimal numbers that are approximate integers).}

\item{options}{Vector of character strings or integers indicating all the possible option values for the \code{data} argument, or \code{NULL}. Numbers of type \code{"double"} are accepted if they have a 0 modulo.}

\item{all_options_in_data}{Single logical value. If \code{TRUE}, all of the options of the \code{options} argument must be present at least once in the \code{data} argument, and nothing else. If \code{FALSE}, some or all of the options must be present in the \code{data} argument, and nothing else. Ignored if the \code{options} argument is \code{NULL}.}

\item{na_contain}{Single logical value. Can the \code{data} argument contain \code{NA}?}

\item{neg_values}{Single logical value. Are negative numeric values authorized? Warning: the default setting is \code{TRUE}, meaning that, in that case, no check is performed for the presence of negative values. The \code{neg_values} argument is activated only when set to \code{FALSE}. In addition, (1) \code{neg_values = FALSE} can only be used when \code{class}, \code{typeof} or \code{mode} arguments are not \code{NULL}, otherwise an error message is returned, (2) the presence of negative values is not checked with \code{neg_values = FALSE} if the tested object is a factor. For the latter, a checking message is returned \code{"OBJECT MUST BE MADE OF NON NEGATIVE VALUES BUT IS A FACTOR"}.}

\item{inf_values}{Single logical value. Are infinite numeric values authorized (\code{Inf} or \code{-Inf})? Same remarks as for the \code{neg_values} argument.}

\item{print}{Single logical value. Print the message if the \code{$problem} output is \code{TRUE}? Warning: set by default to \code{FALSE}, which facilitates the control of the checking message output when using \code{arg_check()} inside functions. See the example section.}

\item{data_name}{Single character string indicating the name of the object to test. If \code{NULL}, use what is assigned to the \code{data} argument for the returned message.}

\item{data_arg}{Single logical value. Is the tested object a function argument? If \code{TRUE} (default), \code{"ARGUMENT"} is written in output messages, otherwise \code{"OBJECT"}.}

\item{safer_check}{Single logical value. Perform some "safer" checks? If \code{TRUE}, checkings are performed before main code running (see \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}. Of note, in \code{arg_check()}, \code{error_text} is also used at the end of the string returned when no problem is detected.}
}
\value{
A list containing:
\itemize{
  \item \code{problem}: logical. Is there any problem detected?
  \item \code{text}: message indicating the details of the problem, or the absence of problem.
  \item \code{object.name}: value of the \code{data_name} argument (i.e., name of the checked object if provided, \code{NULL} otherwise).
}
}
\description{
Check expected values of an argument of functions: class, type, mode, length, restricted values panel, kind of numeric values in addition to the distinction between \code{'integer'} and \code{'double'} (proportion only? Inf values authorized? negative values authorized? Integers of type \code{'double'}?)
}
\details{
If \code{options == NULL}, then at least \code{class} or \code{type} or \code{mode} or \code{length} argument must be non-null.
 
If \code{options} is non-null, then \code{class}, \code{type} and \code{mode} must be \code{NULL}, and \code{length} can be \code{NULL} or specified.
 
The function tests what is written in its arguments, even if what is written is incoherent. For instance, \code{arg_check(data = factor(1), class = "factor", mode = "character")} will return a problem, whatever the object tested in the \code{data} argument, because no object can be class \code{"factor"} and mode \code{"character"} (factors are class \code{"factor"} and mode \code{"numeric"}). Of note, the length of object of class \code{"environment"} is always 0.
 
If the tested object is \code{NULL}, then the function will always return a checking problem.
 
The argument \code{class} can be set to \code{"vector"}. This means that the object is tested for \code{class(data)} returning only \code{"numeric"}, \code{"integer"}, \code{"character"}, \code{"logical"}, \code{"complex"} or \code{"expression"}. Please, use another value of \code{class} (e.g., \code{class = "call"} or \code{class = "list"}) for other types and class of objects.
 
Since R >= 4.0.0, \code{class(matrix())} returns \code{"matrix"} \code{"array"}, and not \code{"matrix"} alone as before. However, use argument \code{class = "matrix"} to check for matrix object (of class \code{"matrix"} \code{"array"} in R >= 4.0.0) and use argument \code{class = "array"} to check for array object (of class \code{"array"} in R >= 4.0.0).
}
\examples{
test <- matrix(1:3)
\dontrun{
# Example that returns an error
saferDev::arg_check(data = test, print = TRUE, class = "vector", mode = "numeric")  # error: THE test ARGUMENT MUST BE CLASS vector
}
saferDev::arg_check(data = test, print = TRUE, class = "matrix", mode = "numeric")
saferDev::arg_check(data = test, print = TRUE, class = "matrix", mode = "numeric", error_text = " using saferDev::arg_check()")
}
\seealso{
\code{\link{match.arg}}.
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

\href{mailto:yushi.han2000@gmail.com}{Yushi Han}

\href{mailto:wanghaiding442@gmail.com}{Haiding Wang}
}
