% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{arg_check}
\alias{arg_check}
\title{arg_check}
\usage{
arg_check(
  data,
  class = NULL,
  typeof = NULL,
  mode = NULL,
  length = NULL,
  prop = FALSE,
  double.as.integer.allowed = FALSE,
  options = NULL,
  all.options.in.data = FALSE,
  na.contain = FALSE,
  neg.values = TRUE,
  inf.values = TRUE,
  print = FALSE,
  data.name = NULL,
  fun.name = NULL
)
}
\arguments{
\item{data}{Object to test.}

\item{class}{Single character string. Either one of the class() result or "vector" or "ggplot2" (i.e., objects of class c("gg", "ggplot")) or NULL. See the warning section below.}

\item{typeof}{Single character string. Either one of the typeof() result or NULL.}

\item{mode}{Single character string. Either one of the mode() result (for non-vector object) or NULL.}

\item{length}{Single numeric value indicating the length of the object. Not considered if NULL.}

\item{prop}{Single logical value. Are the numeric values between 0 and 1 (proportion)? If TRUE, can be used alone, without considering class, etc.}

\item{double.as.integer.allowed}{Single logical value. If TRUE, no error is reported in the cheking message if argument is set to typeof == "integer" or class == "integer", while the reality is typeof == "double" or class == "numeric" but the numbers strictly have zero as modulo (remainder of a division). This means that i <- 1, which is typeof(i) -> "double" is considered as integer with double.as.integer.allowed = TRUE. WARNING: data mod 1 == 0L but not isTRUE(all.equal(data mod 1, 0)) is used here because the argument checks for integers stored as double (does not check for decimal numbers that are approximate integers).}

\item{options}{Vector of character strings or integers indicating all the possible option values for the data argument, or NULL. Numbers of type "double" are accepted if they have a 0 modulo.}

\item{all.options.in.data}{Single logical value. If TRUE, all of the options must be present at least once in the data argument, and nothing else. If FALSE, some or all of the options must be present in the data argument, and nothing else. Ignored if options is NULL.}

\item{na.contain}{Single logical value. Can the data argument contain NA?}

\item{neg.values}{Single logical value. Are negative numeric values authorized? Warning: the default setting is TRUE, meaning that, in that case, no check is performed for the presence of negative values. The neg.values argument is activated only when set to FALSE. In addition, (1) neg.values = FALSE can only be used when class, typeof or mode arguments are not NULL, otherwise return an error message, (2) the presence of negative values is not checked with neg.values = FALSE if the tested object is a factor and the following checking message is returned "OBJECT MUST BE MADE OF NON NEGATIVE VALUES BUT IS A FACTOR".}

\item{inf.values}{Single logical value. Are infinite numeric values authorized (Inf or -Inf)? Identical remarks as for the neg.values argument.}

\item{print}{Single logical value. Print the message if $problem is TRUE? Warning: set by default to FALSE, which facilitates the control of the checking message output when using arg_check() inside functions. See the example section.}

\item{data.name}{Single character string indicating the name of the object to test. If NULL, use what is assigned to the data argument for the returned message.}

\item{fun.name}{Single character string indicating the name of the function checked (i.e., when arg_check() is used to check the arguments of this function). If non-null, the value of fun.name will be added into the message returned by arg_check().}
}
\value{
A list containing:
- §problem: logical. Is there any problem detected?
- §text: message indicating the details of the problem, or the absence of problem.
- §object.name: value of the data.name argument (i.e., name of the checked object if provided, NULL otherwise).
}
\description{
Check the class, type, mode and length of the data argument.

Mainly used to check the arguments of other functions.

Check also other kind of data parameters, is it a proportion? Is it type double but numbers without decimal part?
}
\details{
- If options == NULL, then at least class or type or mode or length argument must be non-null.
 - If options is non-null, then class, type and mode must be NULL, and length can be NULL or specified.
 

WARNINGS

The function tests what is written in its arguments, even if what is written is incoherent. For instance, arg_check(data = factor(1), class = "factor", mode = "character") will return a problem, whatever the object tested in the data argument, because no object can be class "factor" and mode "character" (factors are class "factor" and mode "numeric"). Of note, length of object of class "environment" is always 0.

If the tested object is NULL, then the function will always return a checking problem.

Argument "class" with value "vector" means that the object is tested for class(data) returning only "numeric", "integer", "character", "logical", "complex" or "expression". Please, use another value of class (e.g., class = "call" or class = "list") for other types and class of objects

Since R >= 4.0.0, class(matrix()) returns "matrix" "array", and not "matrix" alone as before. However, use argument class = "matrix" to check for matrix object (of class "matrix" "array" in R >= 4.0.0) and use argument class = "array" to check for array object (of class "array" in R >= 4.0.0).


REQUIRED PACKAGES

None


REQUIRED FUNCTIONS FROM THE cute PACKAGE

None
}
\examples{
test <- matrix(1:3)
# arg_check(data = test, print = TRUE, class = "vector", mode = "numeric")  # commented because this example returns an error
}
