% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.base_op_check.R
\name{.base_op_check}
\alias{.base_op_check}
\title{.base_op_check}
\usage{
.base_op_check(error_text)
}
\arguments{
\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>.". If NULL, converted into "".}
}
\value{
An error message if at least one of the checked operator is present in the R scope, nothing otherwise.
}
\description{
Check if critical operators of R are not present in other packages or in the global env.
Others functions of the R scope can be overwritten because safer functions always use :: when using any function.
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
\keyword{internal}
