% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.extract_all_fun_names.R
\name{.extract_all_fun_names}
\alias{.extract_all_fun_names}
\title{.extract_all_fun_names}
\usage{
.extract_all_fun_names(text, pattern, lib_path, error_text)
}
\arguments{
\item{text}{Single string.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R package are not installed in the default directories because of lack of admin rights.  More precisely, lib_path is passed through the new argument of .libPaths() so that the new library paths are unique(c(new, .Library.site, .Library)). Warning: .libPaths() is restored to the initial paths, after function execution. Ignored if NULL (default) or if the safer_check argument is FALSE: only the pathways specified by the current .libPaths() are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>.". If NULL, converted into "".}

\item{pattern:}{Single string of a perl regex to extract function names.}
}
\value{
List containing:
$string: the function names without parenthesis.
$pos: position of the first character of the function names in the input string
}
\description{
Extract all function names.
}
\details{
- Warning: requires saferDev::arg_check. In main safer functions, in the section "######## check of the required functions from the required packages" add these functions when checking for the presence of saferDev:::.extract_all_fun_names.
}
\examples{
\dontrun{ # Example that shouldn't be run because this is an internal function (not found by devtools::check())
saferDev:::.extract_all_fun_names(text = 'This is a test string with sum()', pattern = "([a-zA-Z]|\\\\.[a-zA-Z._])[a-zA-Z0-9._]*\\\\s*\\\\(", lib_path = NULL, error_text = " INSIDE P1::F1")
}


}
\author{
\href{gael.millot@pasteur.fr}{Gael Millot}
}
\keyword{internal}
