% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.functions_detect.R
\name{.functions_detect}
\alias{.functions_detect}
\title{Functions Detection}
\usage{
.functions_detect(x, skipped_base, arg_user_setting2, lib_path, error_text)
}
\arguments{
\item{x}{Function name to test, written without quotes and brackets.}

\item{skipped_base}{Vector of strings specifying the R basic functions skipped from detection.}

\item{arg_user_setting2}{Argument user settings list.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
A list:
\itemize{
  \item \code{code}: vector of strings of the code of the tested function.
  \item \code{all_basic_funs}: vector or strings of names of all the basic R functions.
  \item \code{fun_names}: list of names of all the functions, not considering \code{base::c("function", "if", "for", "while", "repeat")}. Compartment names indicate the code line number of the functions in \code{code}.
  \item \code{fun_names_pos}: list of position of the first character of each \code{fun_names}. Compartment names indicate the code line number of the functions in \code{code}.
  \item \code{code_line_nb}: vector of integers of the code line numbers of code for each non empty compartment of \code{fun_names} and \code{fun_names_pos}.
  \item \code{internal_fun_names}: vector of string of names of internal functions in the code of the tested function.
  \item \code{arg_user_setting}: list of argument user settings of the tested function.
}
}
\description{
Detect all the functions names inside a function.
}
\details{
Does not check if the functions inside the code exist.

Use the regex pattern \code{"([a-zA-Z]|\\.[a-zA-Z._])[a-zA-Z0-9._]*\\s*\\("} to detect a function in the code.

The result in \code{all_basic_funs} are all the functions in \code{base::c("package:stats", "package:graphics",  "package:grDevices", "package:utils", "package:datasets", "package:methods", "Autoloads", "package:base")}.

Warnings:
\itemize{
  \item Requires \code{saferDev::arg_check}, \code{saferDev:::.extract_all_fun_names}, \code{saferDev:::.has_odd_number_of_quotes}. 
  \item In the safer Backbone section \code{"######## check of the required functions from the required packages"}, add also these 3 functions when checking for the presence of \code{saferDev:::.functions_detect}.
}
}
\examples{
\dontrun{
# Example that shouldn't be run because this is an internal function.
source("C:\\\\Users\\\\gmillot\\\\Documents\\\\Git_projects\\\\safer-r\\\\saferDev\\\\dev\\\\other\\\\test.R")
saferDev:::.functions_detect(x = test, 
arg_user_setting2 = base::list(x = as.name(x = "test")), 
skipped_base = c("function", "if", "for", "while", "repeat", "else"), 
lib_path = NULL, error_text = " INSIDE P1::F1")
}
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

\href{mailto:yushi.han2000@gmail.com}{Yushi Han}

\href{mailto:wanghaiding442@gmail.com}{Haiding Wang}
}
\keyword{internal}
