% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.has_odd_number_of_quotes.R
\name{.has_odd_number_of_quotes}
\alias{.has_odd_number_of_quotes}
\title{.has_odd_number_of_quotes}
\usage{
.has_odd_number_of_quotes(input_string, pattern, lib_path, error_text)
}
\arguments{
\item{input_string}{Single string.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R package are not installed in the default directories because of lack of admin rights.  More precisely, lib_path is passed through the new argument of .libPaths() so that the new library paths are unique(c(new, .Library.site, .Library)). Warning: .libPaths() is restored to the initial paths, after function execution. Ignored if NULL (default) or if the safer_check argument is FALSE: only the pathways specified by the current .libPaths() are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>.". If NULL, converted into "".}

\item{pattern:}{Either '"' or "'".}
}
\value{
TRUE or FALSE.
}
\description{
Determine if a string has a odd number of quotes.
}
\details{
- Consider any present ' (pattern = "'") or " (pattern = '"'). Meaning that, the function does not treat special situation, like escape quotes \", or situation like this "'". All these must be removed from the input_string if necessary, using for instance: input_string <- gsub( x = input_string, pattern = '"\'"', replacement = '   ')
- Warning: requires saferDev::arg_check. In main safer functions, in the section "######## check of the required functions from the required packages" add these functions when checking for the presence of saferDev:::.has_odd_number_of_quotes.
}
\examples{
\dontrun{ # Example that shouldn't be run because this is an internal function
.has_odd_number_of_quotes(input_string = 'This is a "test" string with "even" quotes', pattern = '"', lib_path = NULL, error_text = " INSIDE P1::F1")
}

# importFrom none
}
\author{
Gael Millot <gael.millot@pasteur.fr>
}
\keyword{internal}
