% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.in_quotes_replacement.R
\name{.in_quotes_replacement}
\alias{.in_quotes_replacement}
\title{Replacement Inside Quotes}
\usage{
.in_quotes_replacement(
  string,
  pattern,
  no_regex_pattern,
  replacement,
  perl,
  lib_path,
  error_text
)
}
\arguments{
\item{string}{Single string.}

\item{pattern}{Single string indicating the pattern to detect. Warning: must be very simple pattern, like \code{"\\("}.}

\item{no_regex_pattern}{Same as \code{pattern} without escape characters or list, etc. Used for messages.}

\item{replacement}{Single string for \code{pattern} replacement. Must have the same number of character than \code{no_regex_pattern}. Is not regex.}

\item{perl}{Single logical value. Use Perl regex in \code{pattern} ?}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
A list containing:
\itemize{
  \item \code{string}: The input string with all pattern replaced by the replacement pattern.
  \item \code{pos}: vector of the positions of the 1st character of the replaced pattern. \code{NULL} if no replaced pattern. In that case, \code{string} is identical to the input string.
}
}
\description{
Parse a string and replace any pattern inside simple or double quotes by another replacement pattern and get the position of replacements.
}
\details{
The function works first by replacing in the code: 1) the three consecutive characters \code{'"'} or \code{'"'} by three spaces, 2) \code{"\"'"} and \code{'\'"'} by four spaces and 3) escape quotes like \code{\"} or \code{\'} by two spaces, before looking for the pattern inside quotes.

Warnings:
\itemize{
  \item Must be very simple pattern, like \code{"\\("}.
  \item requires \code{saferDev::arg_check}, \code{saferDev:::.has_odd_number_of_quotes}.
  \item In the safer Backbone section \code{"######## check of the required functions from the required packages"}, add also these 2 functions when checking for the presence of \code{saferDev:::.in_quotes_replacement}.
}
}
\examples{
\dontrun{
# Example that shouldn't be run because this is an internal function.
source("https://raw.githubusercontent.com/safer-r/saferDev/main/dev/other/test.R")
saferDev:::.in_quotes_replacement(string = paste(deparse(test), collapse = ""), 
pattern = "\\\\)", no_regex_pattern = ")", replacement = " ", perl = TRUE, 
lib_path = NULL, error_text = " INSIDE P1::F1")
saferDev:::.in_quotes_replacement(string = 'paste0("IAGE((", paste0(1:3, collapse = " "), "A)B()")', 
pattern = "\\\\)", no_regex_pattern = ")", replacement = " ", perl = TRUE, 
lib_path = NULL, error_text = " INSIDE P1::F1")
}


}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

\href{mailto:yushi.han2000@gmail.com}{Yushi Han}

\href{mailto:wanghaiding442@gmail.com}{Haiding Wang}
}
\keyword{internal}
