% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_.pack_and_function_check.R
\name{.pack_and_function_check}
\alias{.pack_and_function_check}
\title{Package and Function Check}
\usage{
.pack_and_function_check(fun, lib_path, error_text)
}
\arguments{
\item{fun}{Character vector of the names of the required functions, preceded by the name of the package they belong to, separated by a double or triple colon. Example: \code{c("ggplot2::geom_point", "grid::gpar")}.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
An error message if at least one of the checked packages is missing in \code{lib_path}, or if at least one of the checked functions is missing in the required package, nothing otherwise.
}
\description{
Check if 1) required functions are present in required packages and 2) required packages are installed locally.
Simplified version of \code{saferDev::is_function_here()}, used as internal function for the other functions of the saferDev package.
}
\examples{
\dontrun{
# Example that shouldn't be run because this is an internal function.
# this example returns an error
saferDev:::.pack_and_function_check(
    fun = 1, 
    lib_path = NULL, 
    error_text = " INSIDE F1."
) 
saferDev:::.pack_and_function_check(
    fun = "ggplot2::notgood", 
    lib_path = base::.libPaths(), 
    error_text = " INSIDE P1::F1"
) # this example returns an error
saferDev:::.pack_and_function_check(
    fun = c("ggplot2::geom_point", "grid::gpar"), 
    lib_path = base::.libPaths(), 
    error_text = " INSIDE P1::F1"
)  # nothing should happen
}

}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}

Haiding Wang

Yushi Han
}
\keyword{internal}
