% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_function_here.R
\name{is_function_here}
\alias{is_function_here}
\title{is_function_here}
\usage{
is_function_here(fun, lib.path = NULL)
}
\arguments{
\item{fun}{Character vector of the names of the required functions, preceded by the name of the package they belong to and a double colon. Example: c("ggplot2::geom_point", "grid::gpar"). Warning: do not write "()" at the end of the function}

\item{lib.path}{Character vector specifying the absolute pathways of the directories containing the listed packages in the fun argument, if not in the default directories. Ignored if NULL.}
}
\value{
An error message if at least one of the checked packages is missing in lib.path, or if at least one of the checked functions is missing in the required package, nothing otherwise.
}
\description{
Check if required functions are present in installed packages. This controls modifications in of function names package versions.
}
\examples{
# is_function_here(fun = "ggplot2::notgood") # commented because this example returns an error
is_function_here(fun = c("ggplot2::geom_point", "grid::gpar"))
}
