% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package_here.R
\name{is_package_here}
\alias{is_package_here}
\title{is_package_here}
\usage{
is_package_here(req.package, lib.path = NULL)
}
\arguments{
\item{req.package}{Character vector of package names to check.}

\item{lib.path}{Character vector specifying the absolute pathways of the directories containing the listed packages in the req.package argument, if not in the default directories. If NULL, the function checks only in the .libPaths() default R library folders.}
}
\value{
An error message if at least one of the checked packages is missing in lib.path, nothing otherwise.
}
\description{
Check if required packages are installed locally.
}
\examples{
# is_package_here(req.package = "nopackage") # commented because this example returns an error
is_package_here(req.package = "ggplot2")

# is_package_here(req.package = "ggplot2", lib.path = "C:/Users/yhan/AppData/Local/R/win-library/4.3") 
# commented because this example returns an error if the lib.path argument is not an existing directory
}
