% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_colons_check.R
\name{two_colons_check}
\alias{two_colons_check}
\title{two_colons_check}
\usage{
two_colons_check(x, safer_check = TRUE)
}
\arguments{
\item{x}{a function name, written without quotes and brackets.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r) ? If TRUE, two_colons_check() checks before running that 1) the R scope for R operators (like "<-") is not overwritten by another package and 2) functions and related packages used are present in R lybraries. Set to FALSE if two_colons_check() is used inside another "safer" function to avoid pointless multiple checking.}
}
\value{
A message indicating the missing :: or a message saying that everything seems fine.
}
\description{
Verify that all the functions used inside a function are all referenced by their package attribution. For instance: base::mean()
}
\details{
- More precisely, two_colons_check() verifies that all the strings before an opening bracket "(" are preceeded by "::"

- The regex used is: "\\b[a-zA-Z.]{1}[a-zA-Z0-9._]+\\b(?= *\\()"
 
- The following R operators using bracket are not considered: "function", "if", "for", "while" and "repeat"
}
\examples{
two_colons_check(mean)

}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
